Improvements in the code for better readability
Also, fixes in the migration file.
