+++ base/views/organization.py
@@ -28,6 +28,7 @@ from base.forms import OrganizationForm
 from . import layout
 from reference import models as mdlref
 
+
 def organizations(request):
     return layout.render(request, "organizations.html", {'types': mdl.organization.ORGANIZATION_TYPE})
 
@@ -118,7 +119,7 @@ def organization_address_read(request, organization_address_id):
 def organization_address_edit(request, organization_address_id):
     organization_address = mdl.organization_address.find_by_id(organization_address_id)
     organization_id = organization_address.organization.id
-    countries = mdlref.find_all_country()
+    countries = mdlref.country.find_all()
     return layout.render(request, "organization_address_form.html", {'organization_address': organization_address,
                                                                      'organization_id':      organization_id,
                                                                      'countries': countries})
@@ -151,7 +152,7 @@ def organization_address_save(request, organization_address_id):
         organization_address.city = None
 
     if request.POST['country']:
-        organization_address.country = mdlref.find_by_id_country(int(request.POST['country']))
+        organization_address.country = mdlref.country.find_by_id(int(request.POST['country']))
     else:
         organization_address.country = None
