+++ reference/models/country.py
@@ -27,35 +27,6 @@ from django.db import models
 from django.contrib import admin
 
 
-class ContinentAdmin(admin.ModelAdmin):
-    list_display = ('name', 'code',)
-    ordering = ('name',)
-
-
-class Continent(models.Model):
-    code = models.CharField(max_length=2, unique=True)
-    name = models.CharField(max_length=30)
-
-    def __str__(self):
-        return self.name
-
-
-class CurrencyAdmin(admin.ModelAdmin):
-    list_display = ('name', 'code', 'symbol')
-    ordering = ('name',)
-    search_fields = ['code', 'name']
-    fieldsets = ((None, {'fields': ('name', 'code', 'symbol')}),)
-
-
-class Currency(models.Model):
-    name = models.CharField(max_length=80, unique=True)
-    code = models.CharField(max_length=4, blank=True, null=True)
-    symbol = models.CharField(max_length=6, blank=True, null=True)
-
-    def __str__(self):
-        return self.name
-
-
 class CountryAdmin(admin.ModelAdmin):
     list_display = ('name', 'iso_code', 'nationality', 'european_union', 'dialing_code', 'cref_code', 'currency',
                     'continent')
@@ -80,24 +51,9 @@ class Country(models.Model):
         return self.name
 
 
-class LanguageAdmin(admin.ModelAdmin):
-    list_display = ('code', 'name')
-    ordering = ('code',)
-    search_fields = ['code', 'name']
-    fieldsets = ((None, {'fields': ('code', 'name')}),)
+def find_all():
+    return Country.objects.order_by('name')
 
 
-class Language(models.Model):
-    code = models.CharField(max_length=4, unique=True)
-    name = models.CharField(max_length=80, unique=True)
-
-    def __str__(self):
-        return self.name
-
-
-def find_all_country():
-    return Country.objects.all()
-
-
-def find_by_id_country(country_id):
+def find_by_id(country_id):
     return Country.objects.get(pk=country_id)
